Update clean() function

+++ source/functions.2.sh
@@ -66,12 +66,16 @@ function clip() {
 # Clean build artifacts from the current directory.
 function clean() {
     find . -name "*.egg-info" -exec rm -r "{}" +
-    find . -name "dist" -exec rm -r "{}" +
-    find . -name "build" -exec rm -r "{}" +
     find . -name "__pycache__" -exec rm -r "{}" +
     find . -name "*.pyc" -delete
     find . -name ".DS_Store" -delete
-    find . -name "*.class" -delete
+    find . -name "._.DS_Store" -delete
+
+    if [[ "$1" == "-f" || "$1" == "--full" ]]; then
+        find . -name "dist" -exec rm -r "{}" +
+        find . -name "build" -exec rm -r "{}" +
+        find . -name "*.class" -delete
+    fi
 }
 
 # Extract an archive.
